Separated webpack update into its own command
